Update OCP test assertions
Previously, we had a bug in the ocp4e2e suite that caused it to not find
the correct assertion files depending on if there were version-specific
overrides.

  https://github.com/ComplianceAsCode/ocp4e2e/pull/38

This commit updates the test assertions for OCP so that by default they
will find something to check against, which increases the accuracy of
the test coverage.