Move auditing group
Move auditing group so that rules related to auditing are always the
last rules in the built XCCDF benchmark.

This change should prevent situations when a remediation of other rules
from other groups cause a need for changes in audit configuration after
audit has already been remediated. It's a problem mainly for Ansible
Playbooks because they are supposed to be run only once (unlike Bash
remediations and Bash scripts).

This usually happens for the rule audit_rules_privileged_commands which
conflicts with package_sssd_installed. Currently, package_sssd_installed
installs some SUID files which need to be add to audit rules auditing
the privileged commands. But with the current layout, the remediation
for the package_sssd_installed is executed later than remediation for
audit_rules_privileged_commands which means that
audit_rules_privileged_commands starts to fail after sssd is installed.

Unfortunately, moving just a single rule would break the logical
structure and would make the text guides chaotic so we need to move the
whole auditing group.

Fixes: #11752