Add assertion file for RHCOS4 E8 profile tests on 4.13
We're in the process of consolidating all the test assertions for e2e
testing to single files that describe all the results for a particular
scan. This makes it easier to reuse the e2e testing framework for
testing managed offerings, like ROSA.

This commit adds an assertion file for RHCOS4 E8. Test coverage should
be exactly the same as with the assertions in each rule directory.