Restrict the list of accepted shells to avoid user login
There were already long discussions in the past about the cons of using
/bin/false. More recently the topic was also discussed in CIS community
and unanimously agreed to use /sbin/nologin for a user shell when it is
not desired the respective user to interact with shell. This is also in
alignment with STIG recommendations. In addition, is the approach in
all other related rules in the project. Therefore, this OVAL was
updated to the minimal list of shells.