Add new rule file_permissions_var_log_audit_stig
A new rule was created from the existing rule `file_permissions_var_log_audit`
to comply with STIG for Ubuntu 22.04 which requires stricter permissions (0600) on
audit logs compared to CIS (0640).

The new rule was adapted to not only check and remediate the single log
file (log_file in config or default if not present), but all the log
files in the parent directory of the main log file.