Update test scenarios headers
Use the "variables" keyword in the test scenario headers instead of the
"profiles" keyword. These test scenarios aren't a regression test
special to STIG. The profile was set in order to set the XCCDF Value
value. This is legacy usage of the "profiles" keyword and we recommend
using "variables" for this situation.