Extend the Profile class in profile_tool
I had to conduct some tests with variables in profiles and it became
much easier when this class was expanded to also map variables instead of
only rules. It was actually considered when the class was introduced but
there was no concrete demand for this back in the time. Now there is a
demand for assessment of profiles variables.