Fix issue with python executable
The set(Python_ADDITIONAL_VERSIONS 3 2) line was accidentaly moved after
find_package(PythonInterp REQUIRED), impacting in the value of
PYTHON_EXECUTABLE variable. It was noticed in a CI test for CentOS
Stream8 where the "/usr/bin/python3" was still poiting to "/usr/bin/python3.6"
while "/usr/bin/python3.11" was already present and this impacted the
detection of installed Python modules.
Few other improvements in readability were made in this commit.