Improve Ansible remediation for enable_authselect
This is a supportive rule intended to ensure authselect is used in a
system, specially in fresh installations in some specific scenarios
where authselect is not enabled by default. In this case, the preferred
profile is selected. However, in a scenario where a custom profile is
used, even based on the preferred profile, this rule would move again
from the custom profile to the preferred profile. This is usually not a
desired bahavior because may cause inconsistencies between the
profiles since custom profiles are usually there because a requirement
cannot be satisfied by a default profile.