Mark some scenarios as specific to SCE
We introduce a new test scenarios header `check` that allows
to mark test scenarios as specific to a single check engine
type. For example, adding header `check = sce` to a test scenario
marks this test scenario as specific only to SCE. If SCE check
isn't available, such scenario will be skipped. If the `check`
header isn't specified in a test scenario, the test scenario
will work with any check type.

Fixes: #12030