Update OVAL check to use variable value
The regex for the textfilecontent54_object was also updated to consider
a situation where multiple hashing algorithms are specified as
pam_unix.so options. This may cause unexpected behaviors in the system.
If this situation happens, the rule will fail and demand remediation to
make sure that only one hashing algorithm is specified.

Signed-off-by: Marcus Burghardt <maburgha@redhat.com>