Improve the OVAL logic in accounts_password_last_change_is_in_past
Previously, this test was using a hard-coded value to compare the
timestamp of password last change field. This approach works fine in most
cases but proved not robust enough when passwords were changed long time
ago. The new approach uses the current timestamp instead of a hard-coded
value for a more flexible and robust test. Essentially only the
variable_state object was updated, but the ids of some other objects
were renamed for better alignment between objects ids.

Signed-off-by: Marcus Burghardt <maburgha@redhat.com>