CMP-2460: Restrict IdP configurations
The choice of which identity provider to use falls onto the payment
entity. But we already know that htpaswd cannot satisfy PCI-DSS's needs,
and LDAP provider needs to be restrained.

Restrict usage of htpasswd as the identity provider, it cannot provide
user lockout feature required by PCI-DSS.

Do not allow the LDAP provider to transmit clear text passwords.