Update TLS ciphers for ingress controller rule
The following ciphers are all supported with TLS v1.3, but we weren't
checking for them in the OpenShift ingress controller configuration:

  - TLS_AES_128_GCM_SHA256
  - TLS_AES_256_GCM_SHA384
  - TLS_CHACHA20_POLY1305_SHA256
  - TLS_AES_128_CCM_SHA256

This commit updates the regular expression in the rule to check for
those ciphers so the check doesn't fail if OpenShift is using them.

It also add some formatting to the rule so it's consistent with other
TLS-related rules, like for the API server.

The following ciphers were listed in the "old" profile, or insecure, which should
only be used as a last resort for server TLS configuration:

  - AES128-GCM-SHA256
  - AES256-GCM-SHA384
  - DHE-RSA-AES128-GCM-SHA256
  - DHE-RSA-AES256-GCM-SHA384

This commit removes them from the ingress controller rule so that it
fails if a cluster is using these ciphers.

References:
  - https://wiki.mozilla.org/Security/Server_Side_TLS
  - https://docs.openssl.org/1.1.1/man1/ciphers/