Update ciphers in ingress controller remediation
Since we're updating the recommended OCIL, we can also update the
remediation shipped with the content so that it matches. This will allow
users to apply a remediation that updates their TLS ciphers so their
either Recommended or Secure.

This commit has a dependency on a permission change to the operator
cluster role so that it can actually apply the remediation at runtime:

  https://github.com/ComplianceAsCode/compliance-operator/pull/558