Add kubelet tls ingresscontroller rule to CIS benchmarks
This rule was originally written for CIS benchmarks, but somewhere along
the way it was refactored out. This could have been due to a re-indexing
of the controls from the benchmark.

This commit adds the rule back into the CIS profiles so that it's run
with all supports CIS benchmarks.

We should be able to prevent against regressions by including it to the
e2e rule assertion files.