Disable Kickstart remediation in partition_for_boot
In rule partition_for_boot the kickstart remediation conflicts with
`bootprot --kickstart` command which we use in the generated kickstart
by default. This causes problems in some profiles for example RHEL 8
ANSSI where this conflict breaks the installation. We can disable
this remediation for this rule because the creation of the partition
is handled by the `bootprot --kickstart` command.