Build empty OVAL
If we build a thin data stream for a rule that doesn't contain any OVAL
check, the build fails because no OVAL file is generated.  In this
situation, we will generate a minimal OVAL file that doesn't contain any
checks.  Unfortunately, not building the OVAL file at all would be
difficult to be handle by CMake, therefore, we will have a minimal valid
OVAL file as a work around. The artifact doesn't matter because thin
data streams aren't shipped.

Fixes: #12228