Extend mount_option_nodev_nonroot_local_partitions
The OVAL check is extended to read also data directly from the
`/etc/fstab` file. This is useful in environments where the
mount points are not mounted and OVAL partition objects don't
matech. For example, this happens in the Image Builder environment.

Similar to: https://github.com/ComplianceAsCode/content/pull/10200

Resolves: RHEL-45018