Fix chronyd remote server filepath dir regex
Fixes the regular expression that matches 'chrony.d/*.conf' files.

The default path in the project already contains a trailing forward
slash, the rule's filepath was adding an extra slash and causing rule
not to be found.

The rule was likely only working for Debian, as it's product.yml
configures a 'chrony_d_path` without a trailing slash.