Add the references to the selected rules
Change Control.add_references() to iterate over the selected rules, not
the listed rules.

There can be differences on the rules in 'selected' and 'rules'.
When the Control is resolved, the final list of selected rules is in
'selected'.
This is particularly more evident when we are importing other controls.