Update assertions for ingress controller TLS check
We recently incorporated a new rule into the CIS profile that checks
ingress controller TLS configs:

  https://github.com/ComplianceAsCode/content/pull/12220

We added it to the CIS profile, but didn't update the assertions in the
moderate or high profiles, which is causing periodic CI to fail. This
commit adds the assertion to the moderate and high test files so we're
checking it in subsequent CI runs.