Remove assumptions and unnecessary changes
sssd remediation should not create sssd configuration files if they were
not previous created by sys admins. The reason the sssd configuration
files are not created by default is because its parameters may differ
for each site policy, therefore manual intervention is necessary to
ensure the sssd parameters are compliant. The bash remediation was
creating a new file only to satisfy a parameter. This creates
incosistent sssd configuration and makes the sssd service to fail.

Signed-off-by: Marcus Burghardt <maburgha@redhat.com>