Allow rules to have both SCE and OVAL
We will rename the check `variable`.  The `check` variable was shadowed
later when OVAL check is inserted. If a rule had both SCE and OVAL
check, the element referencing the SCE check was overwritten by the
element referencing the OVAL check.  This will allow rules to have both
SCE and OVAL at the same time.