Fix issue with ambiguity of control product and partial match of product names vs product specific controls
When product member is initialised during loading of yaml file it could be ambigously created as a string, when control
is specific for only one product or list when multiple products are specified in the yaml.
The problem with partial match of product names comes later in the add_references method,
 where the current product of the build is matched vs the control product, and the condition used is `product not in self.product`.
In case of list this condition will check if any of the members of the list is exact match to the product.
In case of string though, which is the more common case it will check if the string of the product name, for which we are building is
partially matched (contained) in the self.product.
The issue was found while analysing complaint from a contributor Joel Njanga(@barbarello), while he was trying to add support for al2 platform and it was conflicting with exsting platform al2023