Add the system_with_kernel CPE platform
The system_with_kernel CPE platform will be a common platform
matching all systems that aren't containers.

Matches:
- normal bare-metal systems or VMs
- running bootc (RHEL Image Mode) systems
- bootc (RHEL Image Mode) containers and container images
Does not match:
- classic containers and container images

The matching platforms and non-matching platforms can be easily
distinguished by checking for presence of the kernel RPM package. All
matching platforms have in common that they contain kernel.

Conceptually this platform is an alias of the `package[kernel]`
platform but unfortunately we don't have a way of creating aliases for
CPE platforms, therefore we create a new platform.