Extend `--remove-machine-only` Automatus option
The `system_with_kernel` platform will replace the `machine` platform
in most of the rules. In Automatus we have `--remove-machine-only`
option that intentionally remove the `machine` platform from rules
to force the rule to be applicable in a container back end.
To preserve the ability to test the `machine` rules in a container back
end after we change their platform to `system_with_kernel` we need to
extend the `--remove-machine-only` feature to remove also the
`system_with_kernel` from the temporary data stream.