Unify location of the SCE option
SCE should be built in Ubuntu 20.04 and 22.04 products. However, this
is specified only in the CI workflow description. In previous commit we
have started to build SCE in RHEL 9 and 10. If we would like to start
testing it in CI, we could do it either by changing the CI workflow
description or the build_product script. It would be less complex if we
could unify it in a single place which is the build_product script.