Set "cron_and_at" group platform to system_with_kernel
The rules in "cron_and_at" group should be applicable not only for
classic systems but also for bootable containers because the systems
deployed from bootable containers will be running cron or autodaemon.
Therefore, we are changing the platform from "machine" to the broader
"system_with_kernel" platform. Also, we will remove the forgotten excess
platforms in children of this group.