Set "nfs_and_rpc" group platform to system_with_kernel
The rules in "nfs_and_rpc" group should be applicable not only for
classic systems but also for bootable containers because the systems
deployed from bootable containers will be running nfs_and_rpc.
Therefore, we are changing the platform from "machine" to the broader
"system_with_kernel" platform. Also, we will remove the forgotten excess
platforms in children of this group.