Set "inetd_and_xinetd" group platform to system_with_kernel
The rules in "inetd_and_xinetd" group should be applicable not only for
classic systems but also for bootable containers because the systems
deployed from bootable containers can be configure to run these
services.
Therefore, we are changing the platform from "machine" to the broader
"system_with_kernel" platform. Also, we will remove the forgotten excess
platforms in children of this group.

However, this change is questionable because we don't know if these
packages will ever be part of bootable container images.