Set "ssh" group platform to system_with_kernel
The rules in "ssh" group should be applicable not only for
classic systems but also for bootable containers because the systems
deployed from bootable containers will be running ssh.
Therefore, we are changing the platform from "machine" to the broader
"system_with_kernel" platform. Also, we will remove the forgotten excess
platforms in children of this group.