Set "accounts-physical" group platform to system_with_kernel
The rules in "accounts-physical" group should be applicable not only for
classic systems but also for bootable containers, because the systems
deployed from bootable containers will have physical accounts and will
allow standard user access. Therefore, we are changing the platform
from "machine" to the broader "system_with_kernel" platform.