Set "kernel_build_config" group platform to system_with_kernel
The rules in "kernel_build_config" group should be applicable not only for
classic systems but also for bootable containers, because the systems
deployed from bootable containers will run the kernel that is shipped in
the bootable container image. Therefore, we are changing the platform
from "machine" to the broader "system_with_kernel" platform.

Also, we will remove the forgotten excess platform in a child group.