Set "network-firewalld" group platform to system_with_kernel
The rules in "network-firewalld" group should be applicable not only for
classic systems but also for bootable containers, because the systems
deployed from bootable containers will be normally running firewalld.
Therefore, we are changing the platform from "machine" to the broader
"system_with_kernel" platform.

Also, we will remove the forgotten excess platform in a child group.