Set "network_host_and_router_parameters" group platform to system_with_kernel
The rules in "network_host_and_router_parameters" group should be
applicable not only for classic systems but also for bootable
containers, because the systems deployed from bootable containers will
be able to access network normally and need to have proper network
configuration before connecting. Therefore, we are changing the
platform from "machine" to the broader "system_with_kernel" platform.