Set "network-uncommon" group platform to system_with_kernel
The rules in "network-uncommon" group should be
applicable not only for classic systems but also for bootable
containers, because the systems deployed from bootable containers will
be able to access network normally and need to have the network modules
configure properly after booting. Therefore, we are changing the
platform from "machine" to the broader "system_with_kernel" platform.