Set "selinux" group platform to system_with_kernel
The rules in "selinux" group should be applicable not only for classic
systems but also for bootable containers, because the systems deployed
from bootable containers will be running SELinux and need have a proper
SELinux configuration. Therefore, we are changing the platform from
"machine" to the broader "system_with_kernel" platform.