Set "mcafee_endpoint_security_software" group platform to system_with_kernel
The rules in "mcafee_endpoint_security_software" group should be
applicable not only for classic systems but also for bootable
containers, because the systems deployed from bootable containers will
be typically deployed as end points, so it might make sense to run an
end point security software on these systems.  Therefore, we are
changing the platform from "machine" to the broader "system_with_kernel"
platform.