Set "aide" group platform to system_with_kernel
The rules in "aide" group should be applicable not only for classic
systems but also for bootable containers, because the systems deployed
from bootable containers should need to be configured to run in FIPS
mode. Therefore, we are changing the platform from "machine" to the
broader "system_with_kernel" platform.