Change platform of service enabled or disabled rules
The rules that enable or disable systemd services should
be applicable also in bootable containers because once
a bootable container is booted it can run systemd services.
Therefore we will extend the applicability of the rule to bootable
containers by assigning it the system_with_kernel platform
which is a broader platform than machine.