Change platforms of network group rules
The rules in the "network" group that are marked with the "machine"
platform at this moment should be applicable also to bootable
containers. If a bootable container is booted, the running system will
connect to network and its network configuration needs to be compliant.
We will extend the applicability by changing the CPE platform to a
broader platform "system_with_kernel".