Change platforms of files permissions group rules
The rules in this group that are marked with the "machine" platform at
this moment should be applicable also to bootable containers. If a
bootable container is booted, the file permissions of the files need to
comply with policy requirements.  We will extend the applicability by
changing the CPE platform to a broader platform "system_with_kernel".