Change platforms in rules in restrictions group
The rules in this group that are marked with the "machine" platform at
this moment should be applicable also to bootable containers. If a
bootable container is booted, the system will be running the kernel
shipped in the bootable container image and its configuration will be
applied on that system.  We will extend the rule applicability by
changing the CPE platform to a broader platform "system_with_kernel".