Change platforms in rules in endpoint_security_software group
The rules in this group that are marked with the "machine" platform at
this moment should be applicable also to bootable containers. If a
bootable container is booted, the system will likely act as an endpoint
system and therefore it makes sense to have the endpoint security
software installed and configured in the bootable container image.  We
will extend the rule applicability by changing the CPE platform to a
broader platform "system_with_kernel".