Add SCE check to timer_enabled template
During the build of bootable container images we can't use OVAL check in
rules from the timer_enabled template because the OVAL tests depend on
dbus which isn't available in that environment. Therefore we will use
SCE checks for that environment. The SCE check is similar to the already
existing SCE check for the service_enabled template.