Change platforms of accounts group rules
The rules in the "accounts" group that are marked with the
"machine" platform at this moment should be applicable also
to bootable containers. If a bootable container is booted
the running system will have a quite standard account management.
We will extend the applicability by changing the CPE platform to
a broader platform "system_with_kernel".