Stop SCAP content validation if not necessary
The SCAP content validation is very time consuming. The CTest test cases
that perform the validation takes the most time of the CI runs. However,
the likelihood of catching a serious problem by these tests is low. If
anybody modifies an OVAL, they most likely try to run it in OpenSCAP or
in Automatus tests and these tests would fail because of invalid OVAL.

We will introduce a new CMake option SSG_SCAP_VALIDATION_ENABLED.  If
this option is enable, CTest tests will contain validation of built SCAP
content. The default behavior of the system doesn't change, the option
is set to TRUE by default therefore the validation will be still turned
on by default.

Then, we set this option to OFF in select GitHub Actions CI jobs which
will stop executing SCAP content validation in these jobs.