Add special conditions for bootable containers
This commit adds 2 new Jinja macros: `bootc_build` and
`not_bootc_build`.  These macros define Bash conditional expressions
that are evaluated as true or false if the remediation is performed
during a build of a bootable container image or not performed during a
build of a bootable container image. These macros can be used in Bash
remediation code.  They can be used to control the remediation behavior
in the bootable container build environment.

This commit changes the Bash remediation in rule
`disable_ctrlaltdel_reboot` to demonstrate usefulness of the new macros.