Fix apparmor tests
Tests for ensure_apparmor_enforce were failing because
aa-complain and aa-enforce enabled the rsyslogd profile,
which is disabled by default, thus marking it as an "unconfined
process with a defined profile" and failing the SCE.

Solution is to restart the service when loading the profiles.

Tests for ensure_apparmor_enforce_or_complain were adapted
to not load disabled profiles.