Adapt sysctl template for use in oscap-bootc
OVAL check in sysctl template consists of 2 parts where the first part
checks configuration and second checks the runtime status of the sysctl
option. But, when building a bootable container image we face a problem
that the runtime status doesn't make sense to check and can't be changed
by the remediation. That causes the check after remediation fails and
the rule result is error. Therefore, we need to suppress then runtime part of
the check when building a bootable container image. Also, we shouldn't
attempt to change runtime while building a bootable container image.