Update chronyd rules for bootable containers
The rules used `service_chronyd_enabled` or `service_ntpd_enabled`
in their OVALs to not only check the services configuration but to
also check if the services are enabled.

This approach is unnecessary and it won't work in the bootable
container build environment where systemd services don't run.

Therefore, we update OVALs of these rules and remove the mentioned
service enabled checks and we rather use CPE platforms to make the
rules applicable only in cases when `chrony` or `ntp` packages are
installed. This shouldn't be a big change as if `chrony` or `ntp`
packages are installed their service config files should exist and
in case `chronyd` or `ntpd` services are enabled later they will
already run with the expected configuration.