Add support for /usr/lib/group
If the `nss-altfiles` are installed and `/etc/nsswitch.conf` is
configured to use `nss-altfiles`, the users group can be defined
als in `/usr/lib/group` next to `/etc/group`. The `/usr/lib/group`
is a valid source of group definitions and therefore needs to
be consulted during the check if nsswitch is configured to use
this file. The `nss-altfiles` is often used in bootable containers
base images.