Change platform in dnf-automatic rules
In RHEL Image Mode the automatic updates can't be performed using
dnf-automatic but instead the automatic updates are done by updating to
a new container image version. The automatic updates are handled by the
bootc-fetch-apply-updates.timer and corresponding
bootc-fetch-apply-updates.service. Therefore, the dnf-automatic rules
don't make sense in this environment and we will make them not
applicable there.