Fix sssd_enable_smartcards
If `authselect` doesn't exist and no entry for group + control + module
is present in `/etc/pam.d/system-auth` the Bash remediation inserted
square brackets with backslashes to `system-auth`. The backslashes
shouldn't be inserted there.

The issue manifested when the Bash remediation is used during a build
of a bootable container image.

The solution is to escape the control string only if it's a part of
a regular expression (in `sed` and `grep` commands) but not escape the
control string if it's used in `echo` command.