Remove methods information from classes docstrings
There are cases where this information is useful. Myself can use this
information and can say they are handy. This information causes no harm
and nonody is mandated to use it as well. But there is a conflict of
personal preferences. I wouldn't like to block this PR until we find a
consensus. So I decided to remove it now and keep the discussion in a
separate PR.

Signed-off-by: Marcus Burghardt <maburgha@redhat.com>