Fix hardcoded value for testing oval references
The ctest verify-references-ssg-*-ds.xml failed
after changing oval_feed_url to point to new urls,
due to too many oval references in datastream.

Reason was the the new name of the OVAL feed
was no longer excluded in verify_references.py and
was considered as a regular OVAL file.