Remove Ubuntu oval and remediation for nftables_rules_permanent
Existing ubuntu OVAL checks only whether configuration files are
included in /etc/nftables.conf, and whether they are not empty.
It doesn't check for actual contents, giving a false sense of
correctness when the check passes.

The Ubuntu CIS recommendations require that the content matches
local site policy, which is not possible to automate, thus
removing the audit and remediation.