Update ensure_logrotate_activated for image mode
This rule will be reworked to work well during a bootable container
image build. First, the OVAL will use an internal check for the
logrotate timer instead of extend_definition which uses OVAL
systemdunitproperty test which doesn't work in that environment.
Second, we will installed crontabs package to make sure
that the files required by the bash remediation are present.