Improve rule file_permssions_crontab
This change solves problem of failing rule `file_permissions_crontab`
when the `crontabs` RPM package isn't installed. This situation happens
namely when builidng a CS9 bootable container image because the
base image doesn't contain the crontabs RPM package.

In profiles which contain rules checking permissions on files that are
provided by `crontabs`, we will install the `crontabs` package by
installing the `cronie` package which provides the `cron` service and
has `crontabs` as dependency. This will be achieved by adding the rule
`package_cron_installed` to these profiles.

Notice that the `cronie` package will be installed as a result of rule
`package_cron_installed` which installs either `cron` or `cronie`
based on the product name. This rule is fixed by this commit because
the correct package name for RHEL 9 is `cronie` as well.