Fix sce for ip6tables_rules_for_open_ports
Fixed sce to also consider the protocol of the open port (tcp/udp).
Before, the sce incorrectly passed when an ip6tables rule was defined
for the open port but with the wrong protocol.

The sce was also improved to match ip6tables rules where ports
are defined with --multiport.