Fix audit access rules in ISM_O
The rules `audit_access_failed` and `audit_access_success` fail
after building and booting a CentOS Stream 9 hardened container image
with the `ism_o` profile. The reason is that the remediation fails
to create the files required by these rules because the package `audit`
that provides the directory `/etc/audit/rules.d` where these files
should be created isn't installed by default. The solution is to
install the `audit` package as a part of the profile remediation.