Fix failing file_permissions_crontab
The rule `file_permissions_crontab` fails in a scan performed after
deployment of a CentOS Stream 9 bootable container image hardened
with the PCI-DSS profile. The HTML report shows that the mode of
`/etc/crontab` is `0640` but the rule expects the mode of this
file should be `0600`. The rule passed during the container image
build process because the file `/etc/crontab` didn't exist. The root
cause is that the `cronie` RPM package that provides `/etc/crontab`
is neither present in the CS 9 base image nor it's installed as
a dependency of the PCI-DSS profile. We will fix this problem
by including the rule `package_cron_installed` to the profile
which will install the `cronie` package before `oscap` and then
it will change the `/etc/crontab` mode during remediation.