Fix RHEL 10 ISM profile fails in Image Mode
The rules `audit_access_failed` and `audit_access_success`
fail in the scan of a RHEL 10 Image Mode system hardened with ISM
profiles. The reason is that the profiles don't install the `audit`
RPM package required by these rules to work and at the same time
the `audit` RPM package isn't part of the rhel10-bootc base image.
This is similar to this PR but it fixed only RHEL 9:
https://github.com/ComplianceAsCode/content/pull/12670