disable_ctrlaltdel_burstaction: make sure config file exists.
Starting from RHEL10, the file does not exist by default, it is expected
that the file is created by copying /usr/lib/systemd/system.conf or
creating a new one into /etc/systemd/ or /etc/systemd/system.conf.d/ for
drop-in files.

Just touching the file should be enough for the test scenarios to work
properly. The remediation should still be okay as they will create the
file if is not there and it doesn't need to be a ini file with the
[Manager] key since it loads the /usr/lib/systemd/system.conf file first
and then loads the rest of the files giving priority to drop-in files,
in other words they are loaded last so any configuration there should be
the last to be loaded which should be the one taking place.