Authselect profile minimal is now called local in RHEL10.
Adjust variables and test scenarios to take this into account.
Setting minimal variable to local should make the conversion transparent
for users that have selected this in their profiles. RHEL10 profile will
select minimal as well and it will translate correctly into local
profile which should be the same as minimal. Minimal profile is not
even available as deprecated in RHEL10.

Some test scenarios are not applicable to RHEL10 anymore due to changes
in how pam/authselect interact with the configuration files.